Create compare_sorts.py
The function compares the average times of different sorting algorithms: bubble sort, selection sort and insertion sort. It creates 100 arrays containing 50 integers in the interval [-2000, 2000], which are then sorted by each algorithm, and the sort times are saved in separate files. Finally, the average of the sorting times by each algorithm is calculated and a graph is drawn showing a comparison of the sorting times of successive arrays by the three algorithms.