Create heads_or_tails.py
This is a popular game that involves tossing a coin (1, 2, 5 zl). In each round, a toss of three coins is made. If heads fall on a coin, the value of the coin is added to the balance. The game ends when the balance is equal to or greater than 20 zl. A balance equal to exactly 20 zl means a win, and a larger balance means a loss.